IDEMPIERE-5630 - pr1737 patch
